split vote panel #1381
