+++ src/app/components/common/VotePanel/VotePanelCompact.connect.js
@@ -1,0 +1,4 @@
+import VotePanelConnector from './VotePanelConnector';
+import VotePanelCompact from './VotePanelCompact';
+
+export default VotePanelConnector(VotePanelCompact);
